refactor(spindle-mcp-server): move server implementation to separate file
- Move MCP server implementation from index.ts to server.ts
- Update module exports and imports
- Refactor components, design-doc, design-token, and icon modules
