+++ packages/spindle-mcp-server/src/design-doc.ts
@@ -4,7 +4,7 @@ import fs from 'fs';
 export function getComponentDesignDocTemplate() {
   const templatePath = path.join(
     __dirname,
-    '../../spindle-ui/.scaffdog/design-doc.md',
+    '../assets/spindle-ui/.scaffdog/design-doc.md',
   );
   if (!fs.existsSync(templatePath)) {
     return null;
